ATV_005/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
passlib
python-jose[cryptography]
jinja2
python-dotenv
bcrypt